+++ tests/test.schema.js
@@ -719,10 +719,7 @@ describe('Schema', () => {
       var _expectedDefaultValue = {
         id    : null,
         label : null,
-        obj   : {
-          dishes : [],
-          meals  : []
-        }
+        obj   : null
       };
 
       should(_computed.compilation.main0.level).eql(0);
@@ -771,15 +768,7 @@ describe('Schema', () => {
       var _expectedDefaultValue = {
         id    : null,
         label : null,
-        obj   : {
-          dishes : [],
-          info   : {
-            subinfo : {
-              meals  : [],
-              diners : []
-            }
-          }
-        }
+        obj   : null
       };
 
       // not very good because javascript does not guarantee object attribute order
@@ -1465,6 +1454,38 @@ describe('Schema', () => {
         };
         var _computed = schema.analyzeDescriptor(_objectDescriptor);
         should(_computed.compilation).eql(_expectedCompilationDescriptor);
+        should(_computed.defaultValue).eql(_expectedDefaultValue);
+      });
+
+      it('should set null for an object', () => {
+        var _expectedDefaultValue = {
+          id  : 4,
+          obj : null
+        };
+
+        var _objectDescriptor = {
+          id  : ['<<int>>', 4],
+          obj : ['object', {
+            label : ['string']
+          }]
+        };
+        var _computed = schema.analyzeDescriptor(_objectDescriptor);
+        should(_computed.defaultValue).eql(_expectedDefaultValue);
+      });
+
+      it('should set [] for an array', () => {
+        var _expectedDefaultValue = {
+          id  : 4,
+          obj : []
+        };
+
+        var _objectDescriptor = {
+          id  : ['<<int>>', 4],
+          obj : ['array', {
+            label : ['<<string>>']
+          }]
+        };
+        var _computed = schema.analyzeDescriptor(_objectDescriptor);
         should(_computed.defaultValue).eql(_expectedDefaultValue);
       });
     });